rewrite test_bybit.py to match pytest standards
